requirements.txt
django
djangorestframework
pymysql
waitress
django-filter
django-cors-headers